Consolidated the GAIE and router e2e test files
